requirements.txt
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytorch-grad-cam>=1.5.0
pyyaml>=6.0
tqdm>=4.65.0
